# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Smartbiz Payroll API",
        "description": "API documentation for Smartbiz Payroll",
        "contact": {
            "email": "adisayoga@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://smartbiz-payroll.test/api/v1",
            "description": "Smartbiz Payroll"
        }
    ],
    "paths": {
...
        "/payroll-elements/values/{feedId}": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get one payroll balance feed",
                "operationId": "e4f5df860b909c8fe6bb213dfe6e3d51",
                "parameters": [
                    {
                        "name": "feedId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/payroll-elements/values/{inputValueId}": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get one payroll input value",
                "operationId": "e9edbec59f98051bfcdc124c76f26a6a",
                "parameters": [
                    {
                        "name": "inputValueId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Update payroll input value",
                "operationId": "a45e3493738026f793f56fb34ff9bbf6",
                "parameters": [
                    {
                        "name": "inputValueId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    },
                    {
                        "name": "mode",
                        "in": "query"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Delete payroll input value",
                "operationId": "a152d2c9c8787cca08070fb4275b2c2d",
                "parameters": [
                    {
                        "name": "inputValueId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/payroll-formulas": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get all payroll formula per page",
                "operationId": "b01adb1ad7e829bbfde07b74a42b8506",
                "parameters": [
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "per-page",
                        "in": "query",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filters",
                        "in": "query",
                        "schema": {
                            "type": "{field: string, operator?: string, value: any}[]"
                        }
                    },
                    {
                        "name": "sorts",
                        "in": "query",
                        "schema": {
                            "type": "string[]"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Insert new payroll formula",
                "operationId": "0dfd1fecbc34363673334cae9e21a4ce",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/payroll-formulas/list-cbx": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "List all payroll formula for combob box",
                "operationId": "a2e66b77dd0e08c1c285934fd120666a",
                "parameters": [
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/payroll-formulas/{formulaId}": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get one payroll formula",
                "operationId": "8d594b3880749ddec5834f9d8482b07b",
                "parameters": [
                    {
                        "name": "formulaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Update payroll formula",
                "operationId": "5f90852ee2713d902e70a6ced225a93e",
                "parameters": [
                    {
                        "name": "formulaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    },
                    {
                        "name": "mode",
                        "in": "query"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Delete payroll formula",
                "operationId": "75212e6c2c010d76523225414a3d33c1",
                "parameters": [
                    {
                        "name": "formulaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/payroll-elements/values/{resultId}": {
            "get": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Get one payroll formula result",
                "operationId": "9733cbdf3d48aa563f6a1dcb876e4d5d",
                "parameters": [
                    {
                        "name": "resultId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/payroll-formulas/{formulaId}/results": {
            "post": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Insert new payroll formula result",
                "operationId": "a17e743f4c7f132748849f20b980476c",
                "parameters": [
                    {
                        "name": "formulaId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/payroll-formulas/results/{resultId}": {
            "put": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Update payroll formula result",
                "operationId": "9594aa04dcdecbc09b67d76eb65875fc",
                "parameters": [
                    {
                        "name": "resultId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "effective",
                        "in": "query",
                        "schema": {
                            "type": "date",
                            "format": "yyyy-MM-dd"
                        }
                    },
                    {
                        "name": "mode",
                        "in": "query"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {}
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Compensation"
                ],
                "summary": "Delete payroll formula result",
                "operationId": "67af5018380e897047e9083b97d36bd9",
                "parameters": [
                    {
                        "name": "resultId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format: Bearer [token]",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}